enhance performance for importing to elasticsearch, clean up
